more compiler warning fixes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@5164 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ tools/voxel_grid.cpp
@@ -47,7 +47,7 @@ using namespace pcl;
 using namespace pcl::io;
 using namespace pcl::console;
 
-double      default_leaf_size = 0.01;
+float       default_leaf_size = 0.01f;
 std::string default_field ("z");
 double      default_filter_min = -std::numeric_limits<double>::max ();
 double      default_filter_max = std::numeric_limits<double>::max ();
@@ -156,7 +156,7 @@ main (int argc, char** argv)
   }
   else
   {
-    print_error ("Leaf size must be specified with either 1 or 3 numbers (%lu given). ", (unsigned long) values.size ());
+    print_error ("Leaf size must be specified with either 1 or 3 numbers (%zu given). ", values.size ());
   }
   print_info ("Using a leaf size of: "); print_value ("%f, %f, %f\n", leaf_x, leaf_y, leaf_z);
